Moves more constraint-specific methods to constraint validator class
